fibonana

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,12 +10,12 @@ int i;
 int n = 98;
 unsigned long n1 = 1;
 unsigned long n2 = 2;
-unsigned nth = n1 + n2;
-printf("%d, ", n1);
-printf("%d, ", n2);
+unsigned long nth = n1 + n2;
+printf("%lu, ", n1);
+printf("%lu, ", n2);
 for (i = 0; i < n - 2; i++)
 {
-printf("%d", nth);
+printf("%lu", nth);
 if (i != n - 3)
 {
 printf(", ");
